getAudio/requirements.txt
flask>=3.0
python-dotenv>=1.0
openai-whisper>=20231117
google-genai>=1.0
dashscope>=1.20
requests>=2.28
torch>=2.0
